+++ mq/main/mq-share/src/main/java/com/sun/messaging/jmq/StandardServiceName.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package com.sun.messaging.jmq.jmsserver.service;
+package com.sun.messaging.jmq;
 
 import java.util.Set;
