+++ mq/src/buildant/jarrules.xml
@@ -3,7 +3,7 @@
 
     Copyright (c) 2010, 2020 Oracle and/or its affiliates. All rights reserved.
     Copyright (c) 2020 Payara Services Ltd.
-    Copyright (c) 2021 Contributors to the Eclipse Foundation
+    Copyright (c) 2021, 2024 Contributors to the Eclipse Foundation
 
     This program and the accompanying materials are made available under the
     terms of the Eclipse Public License v. 2.0, which is available at
@@ -207,6 +207,7 @@
 
            	<include name="com/sun/messaging/jmq/DestinationName.class"/>   
            	<include name="com/sun/messaging/jmq/ClientConstants.class"/>   
+           	<include name="com/sun/messaging/jmq/StandardServiceName.class"/>
 
            	<include name="com/sun/messaging/jmq/resources/SharedResources.properties"/>   
            	<include name="com/sun/messaging/jmq/resources/SharedResources.class"/>
